+++ src/main/java/roomescape/theme/service/ThemeService.java
@@ -2,6 +2,7 @@ package roomescape.theme.service;
 
 import java.util.List;
 import org.springframework.stereotype.Service;
+import roomescape.reservation.repository.ReservationRepository;
 import roomescape.theme.domain.Theme;
 import roomescape.theme.repository.ThemeRepository;
 
@@ -9,9 +10,14 @@ import roomescape.theme.repository.ThemeRepository;
 public class ThemeService {
 
     private final ThemeRepository themeRepository;
+    private final ReservationRepository reservationRepository;
 
-    public ThemeService(ThemeRepository themeRepository) {
+    public ThemeService(
+            final ThemeRepository themeRepository,
+            final ReservationRepository reservationRepository
+    ) {
         this.themeRepository = themeRepository;
+        this.reservationRepository = reservationRepository;
     }
 
     public Theme save(final String name, final String description, final String thumbnailUrl) {
@@ -34,6 +40,10 @@ public class ThemeService {
     }
 
     public void deleteById(final long themeId) {
+        if (reservationRepository.existsByThemeId(themeId)) {
+            throw new IllegalArgumentException("[ERROR] 이미 예약된 테마는 삭제할 수 없습니다.");
+        }
+
         themeRepository.deleteById(themeId);
     }
